Backend/otp_verification/requirements.txt
fastapi[all]
twilio
python-dotenv
databases[postgresql]
sqlalchemy
uvicorn
psycopg2-binary==2.9.1
